Added RabbitMqSettings and ServiceSettings

+++ PoliceSupportSystem/Shared.Infrastructure/Settings/RabbitMqSettings.cs
@@ -2,7 +2,7 @@
 
 public record RabbitMqSettings(
     string Host,
-    int Port,
+    ushort Port,
     string? Username = null,
     string? Password = null,
     string? MessageExchange = null,
